+++ lib/rover/vector.rb
@@ -197,7 +197,7 @@ module Rover
       Vector.new(@data.abs)
     end
 
-    [:sqrt, :sin, :cos, :tan, :asin, :acos, :atan, :sinh, :cosh, :tanh, :asinh, :acosh, :atanh].each do |m|
+    [:sqrt, :sin, :cos, :tan, :asin, :acos, :atan, :sinh, :cosh, :tanh, :asinh, :acosh, :atanh, :log, :log2, :log10].each do |m|
       define_method(m) do
         data =
           if @data.is_a?(Numo::SFloat)
@@ -208,6 +208,7 @@ module Rover
         Vector.new(data)
       end
     end
+    alias_method :ln, :log
 
     def each(&block)
       @data.each(&block)