+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.3.3 (unreleased)
 
 - Added `sin`, `cos`, `tan`, `asin`, `acos`, and `atan` methods to vectors
+- Added `sinh`, `cosh`, `tanh`, `asinh`, `acosh`, and `atanh` methods to vectors
 - Added `empty?` method to vectors
 
 ## 0.3.2 (2022-07-10)