+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added `sin`, `cos`, `tan`, `asin`, `acos`, and `atan` methods to vectors
 - Added `sinh`, `cosh`, `tanh`, `asinh`, `acosh`, and `atanh` methods to vectors
 - Added `empty?` method to vectors
+- Added `rename` method to data frames
 
 ## 0.3.2 (2022-07-10)